feat: Extract ImportFilesViewModel startup logic to its own class
