+++ app/src/main/java/com/infomaniak/swisstransfer/ui/screen/newtransfer/ImportFilesViewModelStartup.kt
@@ -1,0 +1,63 @@
+/*
+ * Infomaniak SwissTransfer - Android
+ * Copyright (C) 2025 Infomaniak Network SA
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.infomaniak.swisstransfer.ui.screen.newtransfer
+
+import com.infomaniak.multiplatform_swisstransfer.managers.UploadManager
+import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.channels.Channel.Factory.CONFLATED
+import kotlinx.coroutines.channels.ReceiveChannel
+import kotlinx.coroutines.coroutineScope
+import kotlinx.coroutines.launch
+import javax.inject.Inject
+
+class ImportFilesViewModelStartup @Inject constructor(
+    private val importationFilesManager: ImportationFilesManager,
+    private val newTransferOpenManager: NewTransferOpenManager,
+    private val uploadManager: UploadManager,
+) {
+    private val _openFilePickerEvent: Channel<Unit> = Channel(capacity = CONFLATED)
+    val openFilePickerEvent: ReceiveChannel<Unit> = _openFilePickerEvent
+
+    suspend fun restoreAlreadyImportedFiles() {
+        importationFilesManager.restoreAlreadyImportedFiles()
+    }
+
+    suspend fun handleOpenReasonAsync() {
+        coroutineScope {
+            launch {
+                when (val reason = newTransferOpenManager.readOpenReason()) {
+                    is NewTransferOpenManager.Reason.ExternalShareIncoming -> {
+                        importationFilesManager.importFiles(reason.uris)
+                    }
+                    NewTransferOpenManager.Reason.Other -> {
+                        _openFilePickerEvent.send(Unit)
+                    }
+                }
+            }
+        }
+    }
+
+    suspend fun continuouslyCopyPickedFilesToLocalStorage() {
+        importationFilesManager.continuouslyCopyPickedFilesToLocalStorage()
+    }
+
+    suspend fun removeOldData() {
+        importationFilesManager.removeLocalCopyFolder()
+        uploadManager.removeAllUploadSession()
+    }
+}
